server/pyproject.toml
[project]
name = "trello-server"
version = "0.1.0"
description = "Trello clone – Django REST API"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1",
    "djangorestframework>=3.15",
    "djangorestframework-simplejwt>=5.3",
    "dj-rest-auth[with_social]>=6.0",
    "django-allauth[socialaccount]>=64.0",
    "psycopg[binary]>=3.2",
    "django-cors-headers>=4.5",
    "django-environ>=0.11",
    "gunicorn>=23.0",
    "whitenoise[brotli]>=6.8",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=4.4",
    "watchfiles>=1.0",
    "pytest>=8.0",
    "pytest-django>=4.9",
    "factory-boy>=3.3",
    "ruff>=0.9",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.development"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["apps", "config"]
